Refactor GigsSection into panel components
